requirements.txt
absl-py==0.15.0
astunparse==1.6.3
cachetools==4.2.4
certifi==2021.10.8
charset-normalizer==2.0.7
clang==5.0
dm-tree==0.1.6
flatbuffers==1.12
gast==0.4.0
google-auth==2.3.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.41.0
h5py==3.1.0
idna==3.3
keras==2.6.0
Keras-Preprocessing==1.1.2
Markdown==3.3.4
numpy==1.19.5
oauthlib==3.1.1
onnx==1.10.1
onnx-tf==1.9.0
opt-einsum==3.3.0
Pillow==8.4.0
protobuf==3.19.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.6.0
tensorflow-addons==0.14.0
tensorflow-estimator==2.6.0
tensorflow-model-optimization==0.7.0
termcolor==1.1.0
tf2onnx @ git+https://github.com/onnx/tensorflow-onnx@42e800dc2945e5cadb9df4f09670f2e20eb6d222
torch==1.9.1+cpu
torchaudio==0.9.1
torchvision==0.10.1+cpu
typeguard==2.13.0
typing-extensions==3.7.4.3
urllib3==1.26.7
Werkzeug==2.0.2
wrapt==1.12.1
